The minesweeper app
This is a minesweeper game I created using html, css and JQuery. This
was one of my initial stabs at using JQuery, and overall it's pretty
successful! The current issues with it are that the game doesn't
automatically reveal squares that have no mines adjacent to them if you
select one, you have to do it all yourself.